Dockerfiles/integrated-dev/latest/config/requirements.txt
-i https://pypi.org/simple
jupyter==1.0.0
jupyterlab==2.2.8
jupyterlab-code-formatter==1.3.6
PyYAML==5.3.1
